mistral_cli_tool/pyproject.toml
[project]
name = "mistral-cli-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click",
    "fastmcp>=2.2.5",
    "mistralai",
    "python-dotenv",
    "requests>=2.32.3",
]

[project.scripts]
mistral_cli_tool = "mistral_cli_tool.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
